Add a TEP for Step and Sidecar Workspaces
Prior to this commit Workspace types have been limited to Tasks, TaskRuns, Pipelines and PipelineRuns.
There are known limitations with this:

- Problems with our credentials system are similarly present in our Workspace system: sensitive
data is exposed to every single Step in a Task instead of only those Steps that actually make
use of the data.

- Sidecars do not have a "blessed" path to accessing a Workspace. Instead they must rely on
a hack by adding the $(workspaces.<name>.volume) variable to the `name` field of a `volumeMount`.

The TEP in this commit proposes adding a `workspaces` field to `Steps` and `Sidecars`. The meaning
of this field would be to provide access to that Workspace in either the Step or Sidecar. Explicitly
using a Workspace in a Step would have an additional effect on other Steps: they would be prevented
from accessing the Workspace unless they too were "opted-in" to the Workspace by adding it to their
list of Workspaces.

An example use case for this in Steps is mounting some SSH credentials to perform a clone and then
run tests against that code. If the clone happens in a separate Step from the testing then the SSH
credentials only need to be exposed to the clone. The "untrusted" code in the unit tests does not
need to run with access to the credentials workspace.

An example use case for this in Sidecars is using a Workspace to supply configuration for a mock
API server that runs alongside Steps in a Sidecar and responds with preconfigured responses during
integration tests.