Propose the Workspace Paths feature in a TEP
This commit includes a new TEP for a feature called Workspace Paths.

The Workspace Paths feature provides a way for Task authors to declare
paths on supplied workspace volumes. The paths are divided into those
that the Task expects and those that the Task produces.

Expected paths that the Task declares will be validated before steps
are run, Produced paths that the Task generates will be validated
after steps are run.